refactor: update test files with new hookFormatter import paths
